fix: code review on register screen
