+++ src/components/TextInput/TextInput.tsx
@@ -1,17 +1,10 @@
-import {
-  FocusEvent,
-  forwardRef,
-  ReactElement,
-  ReactNode,
-  useState,
-} from 'react';
+import { FocusEvent, forwardRef, ReactNode, useState } from 'react';
 
 import classNames from 'classnames';
 
-import Icon from '~components/Icon/Icon';
-
 import scss from './TextInput.module.scss';
 
+import { RightIcon } from './TextInput.components';
 import { TInputProps } from './TextInput.types';
 
 function TextInput(props: TInputProps): ReactNode {
@@ -20,14 +13,12 @@ function TextInput(props: TInputProps): ReactNode {
   const containerClass = [scss.container];
   const innerContainerClass = [scss.innerContainer];
   const supportTextClass = [scss.supportText];
-  const rightIconClass = [scss.rightIcon];
 
   const [isFocused, setIsFocused] = useState(false);
 
   if (props.error) {
     containerClass.push(scss.containerError);
     innerContainerClass.push(scss.innerContainerError);
-    rightIconClass.push(scss.rightIconError);
     supportTextClass.push(scss.supportTextError);
     inputClass.push(scss.inputError);
   }
@@ -51,43 +42,6 @@ function TextInput(props: TInputProps): ReactNode {
     setIsFocused(e.target.value.length > 0);
   };
 
-  const renderRightIcon = (): ReactElement | null => {
-    const { error, handleRightIconClick, rightIcon } = props;
-
-    let iconElement;
-
-    if (error)
-      iconElement = (
-        <Icon
-          aria-label="alert-icon"
-          className={classNames(rightIconClass)}
-          icon="alert"
-        />
-      );
-
-    if (rightIcon)
-      iconElement = (
-        <Icon
-          aria-label="input icon"
-          className={classNames(rightIconClass)}
-          icon={rightIcon}
-        />
-      );
-
-    if (!iconElement) return null;
-
-    return handleRightIconClick ? (
-      <button
-        className={classNames(scss.rightIconButton)}
-        onClick={handleRightIconClick}
-      >
-        {iconElement}
-      </button>
-    ) : (
-      iconElement
-    );
-  };
-
   return (
     <div className={classNames(containerClass)}>
       <label className={classNames(labelClass)} htmlFor={props.name}>
@@ -106,7 +60,11 @@ function TextInput(props: TInputProps): ReactNode {
           type={props.type}
           {...props}
         />
-        {renderRightIcon()}
+        <RightIcon
+          error={props.error}
+          handleRightIconClick={props.handleRightIconClick}
+          rightIcon={props.rightIcon}
+        />
       </div>
       <div
         className={classNames(supportTextClass)}
